internal/handlers/recipe_test.go
package handlers

import (
	"encoding/json"
	"log"
	"net/http"
	"net/http/httptest"
	"testing"
)

func TestGetRecipe(t *testing.T) {
	req, err := http.NewRequest("GET", "/recipe/1", nil)
	if err != nil {
		t.Fatal(err)
	}

	rr := httptest.NewRecorder()
	handler := http.HandlerFunc(GetRecipe)
	handler.ServeHTTP(rr, req)
	if status := rr.Code; status != http.StatusOK {
		t.Errorf("handler returned wrong status code: got %v want %v",
			status, http.StatusOK)
	}

	// Check the response body is what we expect.
	result := getRecipeResponse{}
	expected := &getRecipeResponse{
		id:   1,
		name: "test",
	}

	if err := json.NewDecoder(rr.Body).Decode(&result); err != nil {
		log.Fatalln(err)
	}

	if result != *expected {
		t.Errorf("handler returned unexpected body: got %v want %v",
			rr.Body.String(), expected)
	}
}
